// src/models/schemas/User.ts
// Auto-generated by Kadmium. Do not edit manually.
// Use @Kadmium.gen_skip:<name> and @Kadmium.gen_continue:<name> to protect blocks.

import { Model, ModelHookContext, ModelConfig } from "../../model/init";
import { PUBLIC_TYPE_SYMBOL, RELATIONS_SYMBOL } from "../../repo/symbols";
import { ToOneRelation, ToManyRelation } from "../../repo/types/relations";
import { KadmiumRefinementCtx } from "../../validation/types/refinement";
import { Department } from "./Department";
import { Revision } from "../features/Revision";
import { Post } from "./Post";
import { Comment } from "./Comment";

export type UserSensitiveFields = 'password' | 'confirm_password';

export type UserPublic = Omit<User, UserSensitiveFields>;

type __UserRelations = {
	comments: ToManyRelation<Comment>;
	department: ToOneRelation<Department>;
	posts: ToManyRelation<Post>;
	revisions: ToManyRelation<Revision>;
};

export class User extends Model {
	static readonly _collection = "user";
	_meta: "generated-schema" = "generated-schema";
	[PUBLIC_TYPE_SYMBOL]!: UserPublic;
	[RELATIONS_SYMBOL]!: __UserRelations;

	id!: string;
	first_name!: string;
	last_name!: string;
	middle_name!: string | null;
	department_id!: number | null;
	email!: string;
	username!: string;
	password!: string;
	confirm_password!: string;
// @Kadmium.gen_skip:conf
	_conf_: ModelConfig<this> = {
		hooks: {
			beforeCreate: [
				(ctx) => {
					// TODO: implement beforeCreate
				},
			],
			afterCreate: [],
			beforeUpdate: [],
			afterUpdate: [],
			beforeDelete: [],
			afterDelete: [],
			beforeRead: [],
			afterRead: [],
		},
		validation: {
			rules: [
				{
					name: "password_match", refine: (data: this, ctx) => {
						if (data.password !== data.confirm_password) {
							ctx.addIssue(["confirm_password"], "Passwords do not match");
						}
					}
				},
			],
		},
		features: [],
	};
	// @Kadmium.gen_continue:conf
}